feat: move NN converters and layers to separate packages
