+++ src/safeds/ml/nn/converters/_output_conversion_time_series.py
@@ -4,12 +4,13 @@ import sys
 from typing import TYPE_CHECKING, Any
 
 from safeds._utils import _structural_hash
+from safeds.data.labeled.containers import TimeSeriesDataset
+from safeds.data.tabular.containers import Column
+
+from ._output_conversion import OutputConversion
 
 if TYPE_CHECKING:
     from torch import Tensor
-from safeds.data.labeled.containers import TimeSeriesDataset
-from safeds.data.tabular.containers import Column
-from safeds.ml.nn._output_conversion import OutputConversion
 
 
 class OutputConversionTimeSeries(OutputConversion[TimeSeriesDataset, TimeSeriesDataset]):
@@ -78,7 +79,7 @@ class OutputConversionTimeSeries(OutputConversion[TimeSeriesDataset, TimeSeriesD
         window_size: int = kwargs["window_size"]
         forecast_horizon: int = kwargs["forecast_horizon"]
         input_data_table = input_data.to_table()
-        input_data_table = input_data_table.slice_rows(window_size + forecast_horizon)
+        input_data_table = input_data_table.slice_rows(start=window_size + forecast_horizon)
 
         return input_data_table.add_columns(
             [Column(self._prediction_name, output_data.tolist())],
