+++ src/safeds/ml/nn/layers/_forward_layer.py
@@ -3,16 +3,14 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from safeds._config import _init_default_device
+from safeds._utils import _structural_hash
 from safeds._validation import _check_bounds, _ClosedBound
 from safeds.data.image.typing import ImageSize
 
+from ._layer import Layer
+
 if TYPE_CHECKING:
     from torch import Tensor, nn
-
-import sys
-
-from safeds._utils import _structural_hash
-from safeds.ml.nn import Layer
 
 
 def _create_internal_model(input_size: int, output_size: int, activation_function: str) -> nn.Module:
@@ -23,7 +21,7 @@ def _create_internal_model(input_size: int, output_size: int, activation_functio
     class _InternalLayer(nn.Module):
         def __init__(self, input_size: int, output_size: int, activation_function: str):
             super().__init__()
-            self._layer = nn.LSTM(input_size, output_size)
+            self._layer = nn.Linear(input_size, output_size)
             match activation_function:
                 case "sigmoid":
                     self._fn = nn.Sigmoid()
@@ -37,15 +35,15 @@ def _create_internal_model(input_size: int, output_size: int, activation_functio
                     raise ValueError("Unknown Activation Function: " + activation_function)
 
         def forward(self, x: Tensor) -> Tensor:
-            return self._fn(self._layer(x)[0]) if self._fn is not None else self._layer(x)[0]
+            return self._fn(self._layer(x)) if self._fn is not None else self._layer(x)
 
     return _InternalLayer(input_size, output_size, activation_function)
 
 
-class LSTMLayer(Layer):
+class ForwardLayer(Layer):
     def __init__(self, output_size: int, input_size: int | None = None):
         """
-        Create a LSTM Layer.
+        Create a Feed Forward Layer.
 
         Parameters
         ----------
@@ -59,6 +57,7 @@ class LSTMLayer(Layer):
         ValueError
             If input_size < 1
             If output_size < 1
+
         """
         if input_size is not None:
             self._set_input_size(input_size=input_size)
@@ -110,35 +109,29 @@ class LSTMLayer(Layer):
 
     def __hash__(self) -> int:
         """
-        Return a deterministic hash value for this LSTM layer.
+        Return a deterministic hash value for this forward layer.
 
         Returns
         -------
         hash:
             the hash value
         """
-        return _structural_hash(
-            self._input_size,
-            self._output_size,
-        )  # pragma: no cover
+        return _structural_hash(self._input_size, self._output_size)
 
     def __eq__(self, other: object) -> bool:
         """
-        Compare two lstm layer.
-
-        Parameters
-        ----------
-        other:
-            The lstm layer to compare to.
+        Compare two forward layer instances.
 
         Returns
         -------
         equals:
-            Whether the two lstm layer are the same.
+            'True' if input and output size are equal, 'False' otherwise.
         """
-        if not isinstance(other, LSTMLayer):
+        if not isinstance(other, ForwardLayer):
             return NotImplemented
-        return (self is other) or (self._input_size == other._input_size and self._output_size == other._output_size)
+        if self is other:
+            return True
+        return self._input_size == other._input_size and self._output_size == other._output_size
 
     def __sizeof__(self) -> int:
         """
@@ -149,4 +142,6 @@ class LSTMLayer(Layer):
         size:
             Size of this object in bytes.
         """
+        import sys
+
         return sys.getsizeof(self._input_size) + sys.getsizeof(self._output_size)
