+++ src/safeds/ml/nn/layers/_layer.py
