+++ src/safeds/ml/nn/layers/_lstm_layer.py
@@ -1,16 +1,17 @@
 from __future__ import annotations
 
+import sys
 from typing import TYPE_CHECKING, Any
 
 from safeds._config import _init_default_device
+from safeds._utils import _structural_hash
 from safeds._validation import _check_bounds, _ClosedBound
 from safeds.data.image.typing import ImageSize
 
+from ._layer import Layer
+
 if TYPE_CHECKING:
     from torch import Tensor, nn
-
-from safeds._utils import _structural_hash
-from safeds.ml.nn import Layer
 
 
 def _create_internal_model(input_size: int, output_size: int, activation_function: str) -> nn.Module:
@@ -21,7 +22,7 @@ def _create_internal_model(input_size: int, output_size: int, activation_functio
     class _InternalLayer(nn.Module):
         def __init__(self, input_size: int, output_size: int, activation_function: str):
             super().__init__()
-            self._layer = nn.Linear(input_size, output_size)
+            self._layer = nn.LSTM(input_size, output_size)
             match activation_function:
                 case "sigmoid":
                     self._fn = nn.Sigmoid()
@@ -35,15 +36,15 @@ def _create_internal_model(input_size: int, output_size: int, activation_functio
                     raise ValueError("Unknown Activation Function: " + activation_function)
 
         def forward(self, x: Tensor) -> Tensor:
-            return self._fn(self._layer(x)) if self._fn is not None else self._layer(x)
+            return self._fn(self._layer(x)[0]) if self._fn is not None else self._layer(x)[0]
 
     return _InternalLayer(input_size, output_size, activation_function)
 
 
-class ForwardLayer(Layer):
+class LSTMLayer(Layer):
     def __init__(self, output_size: int, input_size: int | None = None):
         """
-        Create a Feed Forward Layer.
+        Create a LSTM Layer.
 
         Parameters
         ----------
@@ -57,7 +58,6 @@ class ForwardLayer(Layer):
         ValueError
             If input_size < 1
             If output_size < 1
-
         """
         if input_size is not None:
             self._set_input_size(input_size=input_size)
@@ -109,29 +109,35 @@ class ForwardLayer(Layer):
 
     def __hash__(self) -> int:
         """
-        Return a deterministic hash value for this forward layer.
+        Return a deterministic hash value for this LSTM layer.
 
         Returns
         -------
         hash:
             the hash value
         """
-        return _structural_hash(self._input_size, self._output_size)
+        return _structural_hash(
+            self._input_size,
+            self._output_size,
+        )  # pragma: no cover
 
     def __eq__(self, other: object) -> bool:
         """
-        Compare two forward layer instances.
+        Compare two lstm layer.
+
+        Parameters
+        ----------
+        other:
+            The lstm layer to compare to.
 
         Returns
         -------
         equals:
-            'True' if input and output size are equal, 'False' otherwise.
+            Whether the two lstm layer are the same.
         """
-        if not isinstance(other, ForwardLayer):
+        if not isinstance(other, LSTMLayer):
             return NotImplemented
-        if self is other:
-            return True
-        return self._input_size == other._input_size and self._output_size == other._output_size
+        return (self is other) or (self._input_size == other._input_size and self._output_size == other._output_size)
 
     def __sizeof__(self) -> int:
         """
@@ -142,6 +148,4 @@ class ForwardLayer(Layer):
         size:
             Size of this object in bytes.
         """
-        import sys
-
         return sys.getsizeof(self._input_size) + sys.getsizeof(self._output_size)
