+++ tests/safeds/ml/nn/converters/test_input_conversion_table.py
@@ -1,5 +1,5 @@
 from safeds.data.labeled.containers import TabularDataset
-from safeds.ml.nn import (
+from safeds.ml.nn.converters import (
     InputConversionTable,
 )
