+++ tests/safeds/ml/nn/test_forward_workflow.py
@@ -3,10 +3,14 @@ from safeds._config import _get_device
 from safeds.data.tabular.containers import Table
 from safeds.data.tabular.transformation import StandardScaler
 from safeds.ml.nn import (
-    ForwardLayer,
-    InputConversionTable,
     NeuralNetworkRegressor,
+)
+from safeds.ml.nn.converters import (
+    InputConversionTable,
     OutputConversionTable,
+)
+from safeds.ml.nn.layers import (
+    ForwardLayer,
 )
 from torch.types import Device
 
@@ -23,12 +27,8 @@ def test_forward_model(device: Device) -> None:
         path=resolve_resource_path(_inflation_path),
     )
     table_1 = table_1.remove_columns(["date"])
-    table_2 = table_1.slice_rows(length=table_1.number_of_rows - 14)
-    table_2 = table_2.add_columns(
-        [
-            table_1.slice_rows(start=14).get_column("value").rename("target"),
-        ]
-    )
+    table_2 = table_1.slice_rows(start=0, length=table_1.number_of_rows - 14)
+    table_2 = table_2.add_columns([(table_1.slice_rows(start=14)).get_column("value").rename("target")])
     train_table, test_table = table_2.split_rows(0.8)
 
     ss = StandardScaler()
